Added below changes,
Renamed html rendering files
Added go test file
Code revamped for readability
Added test in go ci workflow
